requirements.txt
selenium
requests
webdriver-manager
Pillow
